Move class SimpleAudio into the cpp file
